+++ packages/messaging/src/events/user/payloads.ts
@@ -4,6 +4,7 @@ export enum UserEventMessagingType {
   USER_CHANGED = 'user.changed',
   BADGE_CHANGED = 'badge.changed',
   USER_CREATED = 'user.created',
+  USER_DELETED = 'user.deleted',
 }
 
 export interface IBadgePayload {
@@ -28,6 +29,11 @@ export interface IUserPayload {
 }
 
 export interface IUserCreatedPayload {
+  id: string;
+  role: UserRoles;
+}
+
+export interface IUserDeleledPayload {
   id: string;
   role: UserRoles;
 }